Post T2 reattempt to refactor the GoToAprilTag method to the Movement class.
// The initial test to be run on this build will be to align robot during TeleOp and see if holding gamepad1.b causes the robot to align to tag 3.
